+++ shell_integration/par_term_emu_core_rust_shell_integration.bash
@@ -252,8 +252,9 @@ __par_term_emu_preexec() {
     __par_term_emu_last_ret_value="$?"
 
     # OSC 133 ; C - Mark command execution start
+    # $1 contains the command line (provided by bash-preexec)
     par_term_emu_begin_osc
-    printf "133;C;"
+    printf "133;C;%s" "$1"
     par_term_emu_end_osc
 
     # Restore PS1 if we modified it